test(payment): fixed major notes

+++ payment-service/src/main/java/ru/ticketcraft/payment/config/PaymentKafkaProperties.java
@@ -68,7 +68,7 @@ public class PaymentKafkaProperties {
         return replicas;
     }
 
-    public void setReplicas(short replicas) {
+    public void setReplicas(int replicas) {
         this.replicas = replicas;
     }
 }
